Изменения: - расставлены комментарии - создан класс Check_Input для проверки входных данных на пользователя
