+++ app/src/main/AndroidManifest.xml
@@ -10,23 +10,23 @@
         android:supportsRtl="true"
         android:theme="@style/AppTheme">
         <activity
-            android:name=".ui.task_number_6.Activity_Task_06"
+            android:name=".UI.task_number_6.Activity_Task_06"
             android:label="@string/title_activity_task_06"
             android:theme="@style/AppTheme.NoActionBar"></activity>
         <activity
-            android:name=".ui.task_number_5.Activity_Task_05"
+            android:name=".UI.task_number_5.Activity_Task_05"
             android:label="@string/title_activity_task_05"
             android:theme="@style/AppTheme.NoActionBar" />
         <activity
-            android:name=".Activity_Number_1"
+            android:name=".Activities.Activity_Number_1"
             android:label="@string/title_activity_number1"
             android:theme="@style/AppTheme.NoActionBar" />
         <activity
-            android:name=".Activity_Number_2"
+            android:name=".Activities.Activity_Number_2"
             android:label="@string/app_name"
             android:theme="@style/AppTheme.NoActionBar" />
         <activity
-            android:name=".Activity_Main"
+            android:name=".Activities.Activity_Main"
             android:label="@string/app_name"
             android:theme="@style/AppTheme.NoActionBar">
             <intent-filter>
