+++ app/src/main/java/com/example/myapplication/Activities/Activity_Number_1.java
@@ -1,33 +1,28 @@
-package com.example.myapplication.ui.task_number_6;
+package com.example.myapplication.Activities;
 
 import android.os.Bundle;
 
 import com.example.myapplication.R;
-import com.google.android.material.floatingactionbutton.FloatingActionButton;
-import com.google.android.material.snackbar.Snackbar;
 import com.google.android.material.tabs.TabLayout;
 
 import androidx.viewpager.widget.ViewPager;
 import androidx.appcompat.app.AppCompatActivity;
 
-import android.view.Menu;
-import android.view.MenuItem;
-import android.view.View;
+import com.example.myapplication.UI.task_number_1.SectionsPagerAdapter;
 
-import com.example.myapplication.ui.task_number_6.ui.main.SectionsPagerAdapter;
-
-public class Activity_Task_06 extends AppCompatActivity {
+public class Activity_Number_1 extends AppCompatActivity {
 
     @Override
     protected void onCreate(Bundle savedInstanceState) {
         super.onCreate(savedInstanceState);
-        setContentView(R.layout.activity_task_06);
+        // вывставляю layout для отображения
+        setContentView(R.layout.activity_number_1);
+        // объявляем адаптер для того, чтобы отображать наш view_pager
         SectionsPagerAdapter sectionsPagerAdapter = new SectionsPagerAdapter(this, getSupportFragmentManager());
         ViewPager viewPager = findViewById(R.id.view_pager);
         viewPager.setAdapter(sectionsPagerAdapter);
         TabLayout tabs = findViewById(R.id.tabs);
         tabs.setupWithViewPager(viewPager);
-
 
     }
 }
