+++ app/src/main/java/com/example/myapplication/Activities/Activity_Number_2.java
@@ -1,27 +1,27 @@
-package com.example.myapplication;
+package com.example.myapplication.Activities;
 
 import android.os.Bundle;
 
+import com.example.myapplication.R;
 import com.google.android.material.tabs.TabLayout;
 
 import androidx.viewpager.widget.ViewPager;
 import androidx.appcompat.app.AppCompatActivity;
 
-import com.example.myapplication.ui.task_number_1.SectionsPagerAdapter;
+import com.example.myapplication.UI.task_number_2.SectionsPagerAdapter;
 
-public class Activity_Number_1 extends AppCompatActivity {
+public class Activity_Number_2 extends AppCompatActivity {
 
     @Override
     protected void onCreate(Bundle savedInstanceState) {
-        super.onCreate(savedInstanceState);
         // вывставляю layout для отображения
-        setContentView(R.layout.activity_number_1);
+        super.onCreate(savedInstanceState);
+        setContentView(R.layout.activity_number_2);
         // объявляем адаптер для того, чтобы отображать наш view_pager
         SectionsPagerAdapter sectionsPagerAdapter = new SectionsPagerAdapter(this, getSupportFragmentManager());
         ViewPager viewPager = findViewById(R.id.view_pager);
         viewPager.setAdapter(sectionsPagerAdapter);
         TabLayout tabs = findViewById(R.id.tabs);
         tabs.setupWithViewPager(viewPager);
-
     }
 }
