+++ app/src/main/java/com/example/myapplication/UI/PageViewModel.java
@@ -1,9 +1,6 @@
-package com.example.myapplication.ui;
+package com.example.myapplication.UI;
 
-import androidx.arch.core.util.Function;
-import androidx.lifecycle.LiveData;
 import androidx.lifecycle.MutableLiveData;
-import androidx.lifecycle.Transformations;
 import androidx.lifecycle.ViewModel;
 
 public class PageViewModel extends ViewModel {
