+++ app/src/main/java/com/example/myapplication/UI/task_number_1/PlaceholderFragment_Task_3.java
@@ -1,4 +1,4 @@
-package com.example.myapplication.ui.task_number_1;
+package com.example.myapplication.UI.task_number_1;
 
 import android.os.Bundle;
 import android.view.LayoutInflater;
@@ -14,7 +14,7 @@ import androidx.fragment.app.Fragment;
 import androidx.lifecycle.ViewModelProviders;
 
 import com.example.myapplication.R;
-import com.example.myapplication.ui.PageViewModel;
+import com.example.myapplication.UI.PageViewModel;
 
 /**
  * A placeholder fragment containing a simple view.
@@ -59,8 +59,6 @@ public class PlaceholderFragment_Task_3 extends Fragment {
         button = root.findViewById(R.id.button3);
         button.setOnClickListener(oclBtn);
 
-
-
         tAnswer = root.findViewById(R.id.textView3);
         editText = root.findViewById(R.id.editText2);
 
@@ -72,7 +70,7 @@ public class PlaceholderFragment_Task_3 extends Fragment {
         @Override
         public void onClick(View v) {
             switch (v.getId()) {
-                //нажатия на кнопку "SET"e
+                //нажатия на кнопку "SET"
                 case R.id.button3:
                             if (editText.getText().toString().length() != 0) {
                                 tAnswer.setVisibility(View.VISIBLE);
@@ -81,9 +79,6 @@ public class PlaceholderFragment_Task_3 extends Fragment {
                                         "Введите пример!", Toast.LENGTH_SHORT);
                                 toast.show();
                             }
-
-
-
                     break;
 
             }
