+++ app/src/main/java/com/example/myapplication/UI/task_number_6/Activity_Task_06.java
@@ -1,26 +1,27 @@
-package com.example.myapplication;
+package com.example.myapplication.UI.task_number_6;
 
 import android.os.Bundle;
 
+import com.example.myapplication.R;
 import com.google.android.material.tabs.TabLayout;
 
 import androidx.viewpager.widget.ViewPager;
 import androidx.appcompat.app.AppCompatActivity;
 
-import com.example.myapplication.ui.task_number_2.SectionsPagerAdapter;
+import com.example.myapplication.UI.task_number_6.ui.main.SectionsPagerAdapter;
 
-public class Activity_Number_2 extends AppCompatActivity {
+public class Activity_Task_06 extends AppCompatActivity {
 
     @Override
     protected void onCreate(Bundle savedInstanceState) {
-        // вывставляю layout для отображения
         super.onCreate(savedInstanceState);
-        setContentView(R.layout.activity_number_2);
-        // объявляем адаптер для того, чтобы отображать наш view_pager
+        setContentView(R.layout.activity_task_06);
         SectionsPagerAdapter sectionsPagerAdapter = new SectionsPagerAdapter(this, getSupportFragmentManager());
         ViewPager viewPager = findViewById(R.id.view_pager);
         viewPager.setAdapter(sectionsPagerAdapter);
         TabLayout tabs = findViewById(R.id.tabs);
         tabs.setupWithViewPager(viewPager);
+
+
     }
 }
