+++ app/src/main/java/com/example/myapplication/UI/task_number_6/ui/main/SectionsPagerAdapter.java
@@ -1,4 +1,4 @@
-package com.example.myapplication.ui.task_number_1;
+package com.example.myapplication.UI.task_number_6.ui.main;
 
 import android.content.Context;
 
@@ -9,8 +9,7 @@ import androidx.fragment.app.FragmentManager;
 import androidx.fragment.app.FragmentPagerAdapter;
 
 import com.example.myapplication.R;
-import com.example.myapplication.ui.task_number_2.PlaceholderFragment;
-import com.example.myapplication.ui.task_number_2.PlaceholderFragment_2;
+
 
 /**
  * A [FragmentPagerAdapter] that returns a fragment corresponding to
@@ -31,11 +30,9 @@ public class SectionsPagerAdapter extends FragmentPagerAdapter {
     public Fragment getItem(int position) {
         switch (position)
         {
-            case 0: return PlaceholderFragment_Task_1.newInstance(0);
-            case 1: return PlaceholderFragment_Task_2.newInstance(1);
-            case 2: return PlaceholderFragment_Task_3.newInstance(2);
-
-
+            case 0: return PlaceholderFragment.newInstance(0);
+            case 1: return PlaceholderFragment_2.newInstance(1);
+            case 2: return PlaceholderFragment_3.newInstance(2);
         }
         return null;
     }
@@ -48,6 +45,7 @@ public class SectionsPagerAdapter extends FragmentPagerAdapter {
 
     @Override
     public int getCount() {
+        // Show 2 total pages.
         return 3;
     }
 }
