+++ app/src/main/res/layout/activity_number_1.xml
@@ -4,7 +4,7 @@
     xmlns:tools="http://schemas.android.com/tools"
     android:layout_width="match_parent"
     android:layout_height="match_parent"
-    tools:context=".Activity_Number_1">
+    tools:context=".Activities.Activity_Number_1">
 
     <com.google.android.material.appbar.AppBarLayout
         android:layout_width="match_parent"
