+++ app/src/main/res/layout/activity_task_05.xml
@@ -4,7 +4,7 @@
     xmlns:tools="http://schemas.android.com/tools"
     android:layout_width="match_parent"
     android:layout_height="match_parent"
-    tools:context=".ui.task_number_5.Activity_Task_05">
+    tools:context=".UI.task_number_5.Activity_Task_05">
 
     <com.google.android.material.appbar.AppBarLayout
         android:layout_width="match_parent"
