+++ app/src/main/res/layout/fragment_task_0514.xml
@@ -5,7 +5,7 @@
     android:id="@+id/constraintLayout"
     android:layout_width="match_parent"
     android:layout_height="match_parent"
-    tools:context=".ui.task_number_5.ui.main.PlaceholderFragment">
+    tools:context=".UI.task_number_5.ui.main.PlaceholderFragment">
 
     <EditText
         android:id="@+id/editText_lastbit_task14"
