+++ app/src/main/res/layout/fragment_task_1.xml
@@ -5,7 +5,7 @@
     android:id="@+id/constraintLayout"
     android:layout_width="match_parent"
     android:layout_height="match_parent"
-    tools:context=".ui.task_number_1.PlaceholderFragment_Task_1">
+    tools:context=".UI.task_number_1.PlaceholderFragment_Task_1">
 
     <TextView
         android:id="@+id/textView6"
